+++ lib/src/components/material_select/material_select_dropdown_item.dart
@@ -11,6 +11,7 @@ import '../../model/ui/has_renderer.dart';
 import '../../utils/browser/dom_service/dom_service.dart';
 import '../dynamic_component/dynamic_component.dart';
 import '../glyph/glyph.dart';
+import '../material_checkbox/material_checkbox.dart';
 import '../mixins/material_dropdown_base.dart';
 import './activation_handler.dart';
 import './material_select_item.dart';
@@ -54,8 +55,13 @@ import './material_select_item.dart';
           useExisting: MaterialSelectDropdownItemComponent)
     ],
     styleUrls: const ['material_select_dropdown_item.scss.css'],
-    directives: const [DynamicComponent, GlyphComponent, NgIf],
-    templateUrl: 'material_select_dropdown_item.html')
+    directives: const [
+      DynamicComponent,
+      GlyphComponent,
+      MaterialCheckboxComponent,
+      NgIf
+    ],
+    templateUrl: 'material_select_item.html')
 class MaterialSelectDropdownItemComponent extends MaterialSelectItemComponent
     implements OnDestroy {
   MaterialSelectDropdownItemComponent(