+++ lib/src/utils/rate_limit_utils/rate_limit_utils.dart
@@ -7,14 +7,14 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 
 /// A function that can be throttled or debounced.
-typedef UnaryFunction(argument);
-typedef Future DebouncedFunction(argument);
+typedef UnaryFunction<T>(T argument);
+typedef Future DebouncedFunction<T>(T argument);
 typedef Future DebouncedNullaryFunction();
 
 /// A function that rate-limits a delegate function. Helper typedef for
 /// consumers.
-typedef UnaryFunction RateLimitStrategy(
-    UnaryFunction delegate, Duration duration);
+typedef UnaryFunction<T> RateLimitStrategy<T>(
+    UnaryFunction<T> delegate, Duration duration);
 
 /// Returns a wrapper function that, when called with x, executes delegate(x)
 /// [delay] from now iff there is no other call to the wrapper between now and
@@ -23,7 +23,7 @@ typedef UnaryFunction RateLimitStrategy(
 /// The debounced function returns a future that completes when the debounced
 /// function is called. The future is completed with the return results of the
 /// [delegate] function.
-DebouncedFunction debounce(UnaryFunction delegate, Duration delay) {
+DebouncedFunction<T> debounce<T>(UnaryFunction<T> delegate, Duration delay) {
   Timer timer;
   Completer completer;
 
@@ -48,22 +48,22 @@ DebouncedNullaryFunction debounceNullary(void callback(), Duration delay) {
 /// Returns a wrapper function that, when called with x, executes delegate(x)
 /// immediately and prevents further calls to the wrapper from executing
 /// [delegate] until [interval] has elapsed.
-UnaryFunction throttle(UnaryFunction delegate, Duration interval) =>
+UnaryFunction<T> throttle<T>(UnaryFunction<T> delegate, Duration interval) =>
     _throttle(delegate, interval, guaranteeLast: false);
 
 /// Like [throttle], but if the last call to this function is throttled, it will
 /// be executed once the throttling period expires, starting a new throttling
 /// period.
-UnaryFunction throttleGuaranteeLast(
-        UnaryFunction delegate, Duration interval) =>
+UnaryFunction<T> throttleGuaranteeLast<T>(
+        UnaryFunction<T> delegate, Duration interval) =>
     _throttle(delegate, interval, guaranteeLast: true);
 
-UnaryFunction _throttle(UnaryFunction delegate, Duration interval,
+UnaryFunction<T> _throttle<T>(UnaryFunction<T> delegate, Duration interval,
     {@required bool guaranteeLast}) {
   bool onCooldown = false;
   bool hasLastArg = false;
   var lastArg;
-  UnaryFunction self;
+  UnaryFunction<T> self;
   self = (argument) {
     if (!onCooldown) {
       onCooldown = true;